+++ services/firebase/posts.ts
@@ -4,10 +4,7 @@ import { User, Post, PostAll } from '../types'
 
 export const getAllPosts = async (user: Partial<User>): Promise<Post[]> => {
   try {
-    const documents = await firestore()
-      .collection(`discord/${user.userId}/posts`)
-      .orderBy('id', 'desc')
-      .get()
+    const documents = await firestore().collection(`discord/${user.userId}/posts`).orderBy('id', 'desc').get()
 
     const posts: Post[] = []
     for (let index = 0; index < documents.docs.length; index++) {
@@ -40,10 +37,7 @@ export const getAllAllPosts = async (users: User[]): Promise<Post[]> => {
   return posts
 }
 
-export const getPostsByHash = async (
-  user: Partial<User>,
-  hashtag: string
-): Promise<PostAll> => {
+export const getPostsByHash = async (user: Partial<User>, hashtag: string): Promise<PostAll> => {
   const posts: Post[] = []
   try {
     const documents = await firestore()
@@ -71,15 +65,9 @@ export const getPostsByHash = async (
   }
 }
 
-export const getPostById = async (
-  userId: string,
-  id: string
-): Promise<Post | null> => {
+export const getPostById = async (userId: string, id: string): Promise<Post | null> => {
   try {
-    const res = await firestore()
-      .collection(`discord/${userId}/posts`)
-      .doc(id.toLowerCase())
-      .get()
+    const res = await firestore().collection(`discord/${userId}/posts`).doc(id.toLowerCase()).get()
     const data = res.data()
     return data !== undefined ? (data as Post) : null
   } catch (err) {
@@ -88,15 +76,8 @@ export const getPostById = async (
   }
 }
 
-export const updatePost = async (
-  userId: string,
-  id: number,
-  post: Partial<Post>
-): Promise<Post> => {
-  const postDoc = await firestore()
-    .collection(`discord/${userId}/posts`)
-    .doc(String(id))
-    .get()
+export const updatePost = async (userId: string, id: number, post: Partial<Post>): Promise<Post> => {
+  const postDoc = await firestore().collection(`discord/${userId}/posts`).doc(String(id)).get()
   if (!postDoc.exists) {
     const newPost: Post = Object.assign(
       {
@@ -107,10 +88,7 @@ export const updatePost = async (
       },
       post
     )
-    await firestore()
-      .collection(`discord/${userId}/posts`)
-      .doc(String(id))
-      .set(newPost)
+    await firestore().collection(`discord/${userId}/posts`).doc(String(id)).set(newPost)
     return newPost
   }
   await postDoc.ref.update({ ...post, updatedAt: dayjs().toISOString() })
@@ -118,10 +96,7 @@ export const updatePost = async (
 }
 
 export const deletePost = (userId: string, hashtag: string): Promise<any> => {
-  return firestore()
-    .collection(`discord/${userId}/posts`)
-    .doc(hashtag.toLowerCase())
-    .delete()
+  return firestore().collection(`discord/${userId}/posts`).doc(hashtag.toLowerCase()).delete()
 }
 
 export const getLastPost = async (userId: string) => {
